Adds support for additional HTTP methods in API
Updates the CORS configuration in the server to allow "PUT", "GET", and "DELETE" methods in addition to "POST". Also, adds the "address" field to the driver model and updates the driver creation endpoint to include the address field. Lastly, introduces the "estMaintainance" and "lastUpdate" fields to the car model.

Fixes #123